requirements.txt
discord
google-api-python-client
mysql-connector